fix(cli): match AuthorizationItem valueLength to size_t in bb install FFI (#958)
## Summary

Fixes the field type of `value_length` in the hand-written
`AuthorizationItem` FFI struct used by the macOS `bb install`
privileged-symlink flow (`src-tauri/src/commands/cli.rs`).

The Security.framework C definition declares `valueLength` as `size_t`,
which is 8 bytes on 64-bit macOS. Declaring it as `u32` made the Rust
struct layout diverge from the C layout, misaligning the `value` and
`flags` fields that follow it in every `AuthorizationItem` passed to
`AuthorizationCreate` / `AuthorizationCopyRights`. Changing it to
`usize` restores the correct ABI layout.

## Validation

- `just ci` passes (fmt-check, check, lint, typecheck, test,
tauri-check, clippy, build).


---------

+++ src-tauri/src/commands/cli.rs
@@ -326,7 +326,7 @@ mod macos {
     #[repr(C)]
     struct AuthorizationItem {
         name: *const c_char,
-        value_length: u32,
+        value_length: usize,
         value: *mut c_void,
         flags: u32,
     }
